Extract worker and covariance helpers into utils
